RzType introduction
