+++ librz/analysis/analysis.c
@@ -110,7 +110,7 @@ RZ_API RzAnalysis *rz_analysis_new(void) {
 	rz_event_hook(analysis->zign_spaces.event, RZ_SPACE_EVENT_RENAME, zign_rename_for, NULL);
 	rz_analysis_hint_storage_init(analysis);
 	rz_interval_tree_init(&analysis->meta, rz_meta_item_free);
-	analysis->sdb_types = sdb_ns(analysis->sdb, "types", 1);
+	analysis->type = rz_type_new();
 	analysis->sdb_fmts = sdb_ns(analysis->sdb, "spec", 1);
 	analysis->sdb_cc = sdb_ns(analysis->sdb, "cc", 1);
 	analysis->sdb_zigns = sdb_ns(analysis->sdb, "zigns", 1);
@@ -239,39 +239,35 @@ RZ_API bool rz_analysis_set_reg_profile(RzAnalysis *analysis) {
 	return ret;
 }
 
-RZ_API bool rz_analysis_set_triplet(RzAnalysis *analysis, const char *os, const char *arch, int bits) {
+static bool analysis_set_os(RzAnalysis *analysis, const char *os) {
 	rz_return_val_if_fail(analysis, false);
 	if (!os || !*os) {
 		os = RZ_SYS_OS;
 	}
+	free(analysis->os);
+	analysis->os = strdup(os);
+	rz_type_set_os(analysis->type, os);
+	return true;
+}
+
+RZ_API bool rz_analysis_set_triplet(RzAnalysis *analysis, const char *os, const char *arch, int bits) {
+	rz_return_val_if_fail(analysis, false);
 	if (!arch || !*arch) {
 		arch = analysis->cur ? analysis->cur->arch : RZ_SYS_ARCH;
 	}
 	if (bits < 1) {
 		bits = analysis->bits;
 	}
-	free(analysis->os);
-	analysis->os = strdup(os);
+	analysis_set_os(analysis, os);
 	rz_analysis_set_bits(analysis, bits);
 	return rz_analysis_use(analysis, arch);
 }
 
-// copypasta from core/cbin.c
-static void sdb_concat_by_path(Sdb *s, const char *path) {
-	Sdb *db = sdb_new(0, path, 0);
-	sdb_merge(s, db);
-	sdb_close(db);
-	sdb_free(db);
-}
-
 RZ_API bool rz_analysis_set_os(RzAnalysis *analysis, const char *os) {
-	Sdb *types = analysis->sdb_types;
 	const char *dir_prefix = rz_sys_prefix(NULL);
 	const char *dbpath = sdb_fmt(RZ_JOIN_3_PATHS("%s", RZ_SDB_FCNSIGN, "types-%s.sdb"),
 		dir_prefix, os);
-	if (rz_file_exists(dbpath)) {
-		sdb_concat_by_path(types, dbpath);
-	}
+	rz_type_load_sdb(analysis->type, dbpath);
 	return rz_analysis_set_triplet(analysis, os, NULL, -1);
 }
 
@@ -284,6 +280,7 @@ RZ_API bool rz_analysis_set_bits(RzAnalysis *analysis, int bits) {
 	case 64:
 		if (analysis->bits != bits) {
 			analysis->bits = bits;
+			rz_type_set_bits(analysis->type, bits);
 			rz_analysis_set_reg_profile(analysis);
 		}
 		return true;
@@ -298,6 +295,7 @@ RZ_API void rz_analysis_set_cpu(RzAnalysis *analysis, const char *cpu) {
 	if (v != -1) {
 		analysis->pcalign = v;
 	}
+	rz_type_set_cpu(analysis->type, cpu);
 }
 
 RZ_API int rz_analysis_set_big_endian(RzAnalysis *analysis, int bigend) {
@@ -408,7 +406,7 @@ RZ_API void rz_analysis_purge(RzAnalysis *analysis) {
 	rz_analysis_hint_clear(analysis);
 	rz_interval_tree_fini(&analysis->meta);
 	rz_interval_tree_init(&analysis->meta, rz_meta_item_free);
-	sdb_reset(analysis->sdb_types);
+	rz_type_purge(analysis->type);
 	sdb_reset(analysis->sdb_zigns);
 	sdb_reset(analysis->sdb_classes);
 	sdb_reset(analysis->sdb_classes_attrs);
@@ -439,10 +437,10 @@ RZ_API int rz_analysis_archinfo(RzAnalysis *analysis, int query) {
 
 RZ_API bool rz_analysis_noreturn_add(RzAnalysis *analysis, const char *name, ut64 addr) {
 	const char *tmp_name = NULL;
-	Sdb *TDB = analysis->sdb_types;
+	Sdb *NDB = analysis->sdb_noret;
 	char *fnl_name = NULL;
 	if (addr != UT64_MAX) {
-		if (sdb_bool_set(TDB, K_NORET_ADDR(addr), true, 0)) {
+		if (sdb_bool_set(NDB, K_NORET_ADDR(addr), true, 0)) {
 			RzAnalysisFunction *fcn = rz_analysis_get_function_at(analysis, addr);
 			if (fcn) {
 				fcn->is_noreturn = true;
@@ -464,12 +462,12 @@ RZ_API bool rz_analysis_noreturn_add(RzAnalysis *analysis, const char *name, ut6
 			fcn->is_noreturn = true;
 		}
 	}
-	if (rz_type_func_exist(TDB, tmp_name)) {
+	if (rz_type_func_exist(analysis->type, tmp_name)) {
 		fnl_name = strdup(tmp_name);
-	} else if (!(fnl_name = rz_type_func_guess(TDB, (char *)tmp_name))) {
+	} else if (!(fnl_name = rz_type_func_guess(analysis->type, (char *)tmp_name))) {
 		if (addr == UT64_MAX) {
 			if (name) {
-				sdb_bool_set(TDB, K_NORET_FUNC(name), true, 0);
+				sdb_bool_set(NDB, K_NORET_FUNC(name), true, 0);
 			} else {
 				eprintf("Can't find prototype for: %s\n", tmp_name);
 			}
@@ -479,19 +477,19 @@ RZ_API bool rz_analysis_noreturn_add(RzAnalysis *analysis, const char *name, ut6
 		//return false;
 	}
 	if (fnl_name) {
-		sdb_bool_set(TDB, K_NORET_FUNC(fnl_name), true, 0);
+		sdb_bool_set(NDB, K_NORET_FUNC(fnl_name), true, 0);
 		free(fnl_name);
 	}
 	return true;
 }
 
 RZ_API bool rz_analysis_noreturn_drop(RzAnalysis *analysis, const char *expr) {
-	Sdb *TDB = analysis->sdb_types;
+	Sdb *NDB = analysis->sdb_noret;
 	expr = rz_str_trim_head_ro(expr);
 	const char *fcnname = NULL;
 	if (!strncmp(expr, "0x", 2)) {
 		ut64 n = rz_num_math(NULL, expr);
-		sdb_unset(TDB, K_NORET_ADDR(n), 0);
+		sdb_unset(NDB, K_NORET_ADDR(n), 0);
 		RzAnalysisFunction *fcn = rz_analysis_get_fcn_in(analysis, n, -1);
 		if (!fcn) {
 			// eprintf ("can't find function at 0x%"PFMT64x"\n", n);
@@ -501,15 +499,15 @@ RZ_API bool rz_analysis_noreturn_drop(RzAnalysis *analysis, const char *expr) {
 	} else {
 		fcnname = expr;
 	}
-	sdb_unset(TDB, K_NORET_FUNC(fcnname), 0);
+	sdb_unset(NDB, K_NORET_FUNC(fcnname), 0);
 #if 0
 	char *tmp;
 	// unnsecessary checks, imho the noreturn db should be pretty simple to allow forward and custom declarations without having to define the function prototype before
-	if (rz_type_func_exist (TDB, fcnname)) {
-		sdb_unset (TDB, K_NORET_FUNC (fcnname), 0);
+	if (rz_type_func_exist (NDB, fcnname)) {
+		sdb_unset (NDB, K_NORET_FUNC (fcnname), 0);
 		return true;
-	} else if ((tmp = rz_type_func_guess (TDB, (char *)fcnname))) {
-		sdb_unset (TDB, K_NORET_FUNC (fcnname), 0);
+	} else if ((tmp = rz_type_func_guess (NDB, (char *)fcnname))) {
+		sdb_unset (NDB, K_NORET_FUNC (fcnname), 0);
 		free (tmp);
 		return true;
 	}
@@ -519,12 +517,12 @@ RZ_API bool rz_analysis_noreturn_drop(RzAnalysis *analysis, const char *expr) {
 }
 
 static bool rz_analysis_noreturn_at_name(RzAnalysis *analysis, const char *name) {
-	if (sdb_bool_get(analysis->sdb_types, K_NORET_FUNC(name), NULL)) {
+	if (sdb_bool_get(analysis->sdb_noret, K_NORET_FUNC(name), NULL)) {
 		return true;
 	}
-	char *tmp = rz_type_func_guess(analysis->sdb_types, (char *)name);
+	char *tmp = rz_type_func_guess(analysis->type, (char *)name);
 	if (tmp) {
-		if (sdb_bool_get(analysis->sdb_types, K_NORET_FUNC(tmp), NULL)) {
+		if (sdb_bool_get(analysis->sdb_noret, K_NORET_FUNC(tmp), NULL)) {
 			free(tmp);
 			return true;
 		}
@@ -537,7 +535,7 @@ static bool rz_analysis_noreturn_at_name(RzAnalysis *analysis, const char *name)
 }
 
 RZ_API bool rz_analysis_noreturn_at_addr(RzAnalysis *analysis, ut64 addr) {
-	return sdb_bool_get(analysis->sdb_types, K_NORET_ADDR(addr), NULL);
+	return sdb_bool_get(analysis->sdb_noret, K_NORET_ADDR(addr), NULL);
 }
 
 static bool noreturn_recurse(RzAnalysis *analysis, ut64 addr) {
