+++ librz/core/cmd_meta.c
@@ -554,7 +554,7 @@ static int cmd_meta_others(RzCore *core, const char *input) {
 					if (p) {
 						p = (char *)rz_str_trim_head_ro(p);
 						if (*p == '.') {
-							const char *realformat = rz_print_format_byname(core->print, p + 1);
+							const char *realformat = rz_type_format_byname(core->print, p + 1);
 							if (realformat) {
 								p = (char *)realformat;
 							} else {
@@ -563,17 +563,17 @@ static int cmd_meta_others(RzCore *core, const char *input) {
 							}
 						}
 						if (n < 1) {
-							n = rz_print_format_struct_size(core->print, p, 0, 0);
+							n = rz_type_format_struct_size(core->print, p, 0, 0);
 							if (n < 1) {
 								eprintf("Warning: Cannot resolve struct size for '%s'\n", p);
 								n = 32; //
 							}
 						}
-						//make sure we do not overflow on rz_print_format
+						//make sure we do not overflow on rz_type_format
 						if (n > core->blocksize) {
 							n = core->blocksize;
 						}
-						int r = rz_print_format(core->print, addr, core->block,
+						int r = rz_type_format(core->print, addr, core->block,
 							n, p, 0, NULL, NULL);
 						if (r < 0) {
 							n = -1;
