+++ librz/core/meson.build
@@ -97,6 +97,7 @@ rz_core_deps = [
   rz_egg_dep,
   rz_search_dep,
   rz_analysis_dep,
+  rz_type_dep,
   rz_debug_dep,
   rz_config_dep,
   rz_bin_dep,
@@ -138,6 +139,7 @@ pkgconfig_mod.generate(
     'rz_cons',
     'rz_analysis',
     'rz_socket',
+	'rz_type',
     'rz_io',
     'rz_lang',
     'rz_hash',
