+++ librz/meson.build
@@ -381,7 +381,6 @@ rz_util_files = [
   'include/rz_util/rz_bitmap.h',
   'include/rz_util/rz_buf.h',
   'include/rz_util/rz_cache.h',
-  'include/rz_util/rz_ctypes.h',
   'include/rz_util/rz_debruijn.h',
   'include/rz_util/rz_event.h',
   'include/rz_util/rz_file.h',
