+++ librz/type/base.c
@@ -1,0 +1,566 @@
+// SPDX-FileCopyrightText: 2021 Anton Kochkov <anton.kochkov@gmail.com>
+// SPDX-License-Identifier: LGPL-3.0-only
+
+#include <rz_util.h>
+#include <rz_type.h>
+#include <string.h>
+#include <sdb.h>
+
+#include "type_internal.h"
+
+RZ_API void rz_type_base_enum_case_free(void *e, void *user) {
+	(void)user;
+	RzTypeEnumCase *cas = e;
+	free((char *)cas->name);
+}
+
+RZ_API void rz_type_base_struct_member_free(void *e, void *user) {
+	(void)user;
+	RzTypeStructMember *member = e;
+	free((char *)member->name);
+	free((char *)member->type);
+}
+
+RZ_API void rz_type_base_union_member_free(void *e, void *user) {
+	(void)user;
+	RzTypeUnionMember *member = e;
+	free((char *)member->name);
+	free((char *)member->type);
+}
+
+static char *get_type_data(Sdb *sdb_types, const char *type, const char *sname) {
+	char *key = rz_str_newf("%s.%s", type, sname);
+	if (!key) {
+		return NULL;
+	}
+	char *members = sdb_get(sdb_types, key, NULL);
+	free(key);
+	return members;
+}
+
+static RzBaseType *get_enum_type(RzType *type, const char *sname) {
+	rz_return_val_if_fail(type && sname, NULL);
+
+	RzBaseType *base_type = rz_type_base_type_new(RZ_BASE_TYPE_KIND_ENUM);
+	if (!base_type) {
+		return NULL;
+	}
+
+	char *members = get_type_data(type->sdb_types, "enum", sname);
+	if (!members) {
+		goto error;
+	}
+
+	RzVector *cases = &base_type->enum_data.cases;
+	if (!rz_vector_reserve(cases, (size_t)sdb_alen(members))) {
+		goto error;
+	}
+
+	char *cur;
+	sdb_aforeach(cur, members) {
+		char *val_key = rz_str_newf("enum.%s.%s", sname, cur);
+		if (!val_key) {
+			goto error;
+		}
+		const char *value = sdb_const_get(type->sdb_types, val_key, NULL);
+		free(val_key);
+
+		if (!value) { // if nothing is found, ret NULL
+			goto error;
+		}
+
+		RzTypeEnumCase cas = { .name = strdup(cur), .val = strtol(value, NULL, 16) };
+
+		void *element = rz_vector_push(cases, &cas); // returns null if no space available
+		if (!element) {
+			goto error;
+		}
+
+		sdb_aforeach_next(cur);
+	}
+	free(members);
+
+	return base_type;
+
+error:
+	free(members);
+	rz_type_base_type_free(base_type);
+	return NULL;
+}
+
+static RzBaseType *get_struct_type(RzType *type, const char *sname) {
+	rz_return_val_if_fail(type && sname, NULL);
+
+	RzBaseType *base_type = rz_type_base_type_new(RZ_BASE_TYPE_KIND_STRUCT);
+	if (!base_type) {
+		return NULL;
+	}
+
+	char *sdb_members = get_type_data(type->sdb_types, "struct", sname);
+	if (!sdb_members) {
+		goto error;
+	}
+
+	RzVector *members = &base_type->struct_data.members;
+	if (!rz_vector_reserve(members, (size_t)sdb_alen(sdb_members))) {
+		goto error;
+	}
+
+	char *cur;
+	sdb_aforeach(cur, sdb_members) {
+		char *type_key = rz_str_newf("struct.%s.%s", sname, cur);
+		if (!type_key) {
+			goto error;
+		}
+		char *values = sdb_get(type->sdb_types, type_key, NULL);
+		free(type_key);
+
+		if (!values) {
+			goto error;
+		}
+		char *offset = NULL;
+		char *type = sdb_anext(values, &offset);
+		if (!offset) { // offset is missing, malformed state
+			free(values);
+			goto error;
+		}
+		offset = sdb_anext(offset, NULL);
+		RzTypeStructMember cas = {
+			.name = strdup(cur),
+			.type = strdup(type),
+			.offset = strtol(offset, NULL, 10)
+		};
+
+		free(values);
+
+		void *element = rz_vector_push(members, &cas); // returns null if no space available
+		if (!element) {
+			goto error;
+		}
+
+		sdb_aforeach_next(cur);
+	}
+	free(sdb_members);
+
+	return base_type;
+
+error:
+	rz_type_base_type_free(base_type);
+	free(sdb_members);
+	return NULL;
+}
+
+static RzBaseType *get_union_type(RzType *type, const char *sname) {
+	rz_return_val_if_fail(type && sname, NULL);
+
+	RzBaseType *base_type = rz_type_base_type_new(RZ_BASE_TYPE_KIND_UNION);
+	if (!base_type) {
+		return NULL;
+	}
+
+	char *sdb_members = get_type_data(type->sdb_types, "union", sname);
+	if (!sdb_members) {
+		goto error;
+	}
+
+	RzVector *members = &base_type->union_data.members;
+	if (!rz_vector_reserve(members, (size_t)sdb_alen(sdb_members))) {
+		goto error;
+	}
+
+	char *cur;
+	sdb_aforeach(cur, sdb_members) {
+		char *type_key = rz_str_newf("union.%s.%s", sname, cur);
+		if (!type_key) {
+			goto error;
+		}
+		char *values = sdb_get(type->sdb_types, type_key, NULL);
+		free(type_key);
+
+		if (!values) {
+			goto error;
+		}
+		char *value = sdb_anext(values, NULL);
+		RzTypeUnionMember cas = { .name = strdup(cur), .type = strdup(value) };
+		free(values);
+
+		void *element = rz_vector_push(members, &cas); // returns null if no space available
+		if (!element) {
+			goto error;
+		}
+
+		sdb_aforeach_next(cur);
+	}
+	free(sdb_members);
+
+	return base_type;
+
+error:
+	rz_type_base_type_free(base_type);
+	free(sdb_members);
+	return NULL;
+}
+
+static RzBaseType *get_typedef_type(RzType *type, const char *sname) {
+	rz_return_val_if_fail(type && RZ_STR_ISNOTEMPTY(sname), NULL);
+
+	RzBaseType *base_type = rz_type_base_type_new(RZ_BASE_TYPE_KIND_TYPEDEF);
+	if (!base_type) {
+		return NULL;
+	}
+
+	base_type->type = get_type_data(type->sdb_types, "typedef", sname);
+	if (!base_type->type) {
+		goto error;
+	}
+	return base_type;
+
+error:
+	rz_type_base_type_free(base_type);
+	return NULL;
+}
+
+static RzBaseType *get_atomic_type(RzType *type, const char *sname) {
+	rz_return_val_if_fail(type && RZ_STR_ISNOTEMPTY(sname), NULL);
+
+	RzBaseType *base_type = rz_type_base_type_new(RZ_BASE_TYPE_KIND_ATOMIC);
+	if (!base_type) {
+		return NULL;
+	}
+
+	base_type->type = get_type_data(type->sdb_types, "type", sname);
+	if (!base_type->type) {
+		goto error;
+	}
+
+	RzStrBuf key;
+	base_type->size = sdb_num_get(type->sdb_types, rz_strbuf_initf(&key, "type.%s.size", sname), 0);
+	rz_strbuf_fini(&key);
+
+	return base_type;
+
+error:
+	rz_type_base_type_free(base_type);
+	return NULL;
+}
+
+// returns NULL if name is not found or any failure happened
+RZ_API RzBaseType *rz_type_get_base_type(RzType *t, const char *name) {
+	rz_return_val_if_fail(t && name, NULL);
+
+	char *sname = rz_str_sanitize_sdb_key(name);
+	const char *type = sdb_const_get(t->sdb_types, sname, NULL);
+	if (!type) {
+		free(sname);
+		return NULL;
+	}
+
+	RzBaseType *base_type = NULL;
+	if (!strcmp(type, "struct")) {
+		base_type = get_struct_type(t, sname);
+	} else if (!strcmp(type, "enum")) {
+		base_type = get_enum_type(t, sname);
+	} else if (!strcmp(type, "union")) {
+		base_type = get_union_type(t, sname);
+	} else if (!strcmp(type, "typedef")) {
+		base_type = get_typedef_type(t, sname);
+	} else if (!strcmp(type, "type")) {
+		base_type = get_atomic_type(t, sname);
+	}
+
+	if (base_type) {
+		base_type->name = sname;
+	} else {
+		free(sname);
+	}
+
+	return base_type;
+}
+
+static void save_struct(const RzType *t, const RzBaseType *type) {
+	rz_return_if_fail(t && type && type->name && type->kind == RZ_BASE_TYPE_KIND_STRUCT);
+	char *kind = "struct";
+	/*
+		C:
+		struct name {type param1; type param2; type paramN;};
+		Sdb:
+		name=struct
+		struct.name=param1,param2,paramN
+		struct.name.param1=type,0,0
+		struct.name.param2=type,4,0
+		struct.name.paramN=type,8,0
+	*/
+	char *sname = rz_str_sanitize_sdb_key(type->name);
+	// name=struct
+	sdb_set(t->sdb_types, sname, kind, 0);
+
+	RzStrBuf arglist;
+	RzStrBuf param_key;
+	RzStrBuf param_val;
+	rz_strbuf_init(&arglist);
+	rz_strbuf_init(&param_key);
+	rz_strbuf_init(&param_val);
+
+	int i = 0;
+	RzTypeStructMember *member;
+	rz_vector_foreach(&type->struct_data.members, member) {
+		// struct.name.param=type,offset,argsize
+		char *member_sname = rz_str_sanitize_sdb_key(member->name);
+		sdb_set(t->sdb_types,
+			rz_strbuf_setf(&param_key, "%s.%s.%s", kind, sname, member_sname),
+			rz_strbuf_setf(&param_val, "%s,%zu,%u", member->type, member->offset, 0), 0ULL);
+		free(member_sname);
+
+		rz_strbuf_appendf(&arglist, (i++ == 0) ? "%s" : ",%s", member->name);
+	}
+	// struct.name=param1,param2,paramN
+	char *key = rz_str_newf("%s.%s", kind, sname);
+	sdb_set(t->sdb_types, key, rz_strbuf_get(&arglist), 0);
+	free(key);
+
+	free(sname);
+
+	rz_strbuf_fini(&arglist);
+	rz_strbuf_fini(&param_key);
+	rz_strbuf_fini(&param_val);
+}
+
+static void save_union(const RzType *t, const RzBaseType *type) {
+	rz_return_if_fail(t && type && type->name && type->kind == RZ_BASE_TYPE_KIND_UNION);
+	const char *kind = "union";
+	/*
+	C:
+	union name {type param1; type param2; type paramN;};
+	Sdb:
+	name=union
+	union.name=param1,param2,paramN
+	union.name.param1=type,0,0
+	union.name.param2=type,0,0
+	union.name.paramN=type,0,0
+	*/
+	char *sname = rz_str_sanitize_sdb_key(type->name);
+	// name=union
+	sdb_set(t->sdb_types, sname, kind, 0);
+
+	RzStrBuf arglist;
+	RzStrBuf param_key;
+	RzStrBuf param_val;
+	rz_strbuf_init(&arglist);
+	rz_strbuf_init(&param_key);
+	rz_strbuf_init(&param_val);
+
+	int i = 0;
+	RzTypeUnionMember *member;
+	rz_vector_foreach(&type->union_data.members, member) {
+		// union.name.arg1=type,offset,argsize
+		char *member_sname = rz_str_sanitize_sdb_key(member->name);
+		sdb_set(t->sdb_types,
+			rz_strbuf_setf(&param_key, "%s.%s.%s", kind, sname, member_sname),
+			rz_strbuf_setf(&param_val, "%s,%zu,%u", member->type, member->offset, 0), 0ULL);
+		free(member_sname);
+
+		rz_strbuf_appendf(&arglist, (i++ == 0) ? "%s" : ",%s", member->name);
+	}
+	// union.name=arg1,arg2,argN
+	char *key = rz_str_newf("%s.%s", kind, sname);
+	sdb_set(t->sdb_types, key, rz_strbuf_get(&arglist), 0);
+	free(key);
+
+	free(sname);
+
+	rz_strbuf_fini(&arglist);
+	rz_strbuf_fini(&param_key);
+	rz_strbuf_fini(&param_val);
+}
+
+static void save_enum(const RzType *t, const RzBaseType *type) {
+	rz_return_if_fail(t && type && type->name && type->kind == RZ_BASE_TYPE_KIND_ENUM);
+	/*
+		C:
+			enum name {case1 = 1, case2 = 2, caseN = 3};
+		Sdb:
+		name=enum
+		enum.name=arg1,arg2,argN
+		enum.MyEnum.0x1=arg1
+		enum.MyEnum.0x3=arg2
+		enum.MyEnum.0x63=argN
+		enum.MyEnum.arg1=0x1
+		enum.MyEnum.arg2=0x63
+		enum.MyEnum.argN=0x3
+	*/
+	char *sname = rz_str_sanitize_sdb_key(type->name);
+	sdb_set(t->sdb_types, sname, "enum", 0);
+
+	RzStrBuf arglist;
+	RzStrBuf param_key;
+	RzStrBuf param_val;
+	rz_strbuf_init(&arglist);
+	rz_strbuf_init(&param_key);
+	rz_strbuf_init(&param_val);
+
+	int i = 0;
+	RzTypeEnumCase *cas;
+	rz_vector_foreach(&type->enum_data.cases, cas) {
+		// enum.name.arg1=type,offset,???
+		char *case_sname = rz_str_sanitize_sdb_key(cas->name);
+		sdb_set(t->sdb_types,
+			rz_strbuf_setf(&param_key, "enum.%s.%s", sname, case_sname),
+			rz_strbuf_setf(&param_val, "0x%" PFMT32x "", cas->val), 0);
+
+		sdb_set(t->sdb_types,
+			rz_strbuf_setf(&param_key, "enum.%s.0x%" PFMT32x "", sname, cas->val),
+			case_sname, 0);
+		free(case_sname);
+
+		rz_strbuf_appendf(&arglist, (i++ == 0) ? "%s" : ",%s", cas->name);
+	}
+	// enum.name=arg1,arg2,argN
+	char *key = rz_str_newf("enum.%s", sname);
+	sdb_set(t->sdb_types, key, rz_strbuf_get(&arglist), 0);
+	free(key);
+
+	free(sname);
+
+	rz_strbuf_fini(&arglist);
+	rz_strbuf_fini(&param_key);
+	rz_strbuf_fini(&param_val);
+}
+
+static void save_atomic_type(const RzType *t, const RzBaseType *type) {
+	rz_return_if_fail(t && type && type->name && type->kind == RZ_BASE_TYPE_KIND_ATOMIC);
+	/*
+		C: (cannot define a custom atomic type)
+		Sdb:
+		char=type
+		type.char=c
+		type.char.size=8
+	*/
+	char *sname = rz_str_sanitize_sdb_key(type->name);
+	sdb_set(t->sdb_types, sname, "type", 0);
+
+	RzStrBuf key;
+	RzStrBuf val;
+	rz_strbuf_init(&key);
+	rz_strbuf_init(&val);
+
+	sdb_set(t->sdb_types,
+		rz_strbuf_setf(&key, "type.%s.size", sname),
+		rz_strbuf_setf(&val, "%" PFMT64u "", type->size), 0);
+
+	sdb_set(t->sdb_types,
+		rz_strbuf_setf(&key, "type.%s", sname),
+		type->type, 0);
+
+	free(sname);
+
+	rz_strbuf_fini(&key);
+	rz_strbuf_fini(&val);
+}
+
+static void save_typedef(const RzType *t, const RzBaseType *type) {
+	rz_return_if_fail(t && type && type->name && type->kind == RZ_BASE_TYPE_KIND_TYPEDEF);
+	/*
+		C:
+		typedef char byte;
+		Sdb:
+		byte=typedef
+		typedef.byte=char
+	*/
+	char *sname = rz_str_sanitize_sdb_key(type->name);
+	sdb_set(t->sdb_types, sname, "typedef", 0);
+
+	RzStrBuf key;
+	RzStrBuf val;
+	rz_strbuf_init(&key);
+	rz_strbuf_init(&val);
+
+	sdb_set(t->sdb_types,
+		rz_strbuf_setf(&key, "typedef.%s", sname),
+		rz_strbuf_setf(&val, "%s", type->type), 0);
+
+	free(sname);
+
+	rz_strbuf_fini(&key);
+	rz_strbuf_fini(&val);
+}
+
+RZ_API void rz_type_base_type_free(RzBaseType *type) {
+	rz_return_if_fail(type);
+	RZ_FREE(type->name);
+	RZ_FREE(type->type);
+
+	switch (type->kind) {
+	case RZ_BASE_TYPE_KIND_STRUCT:
+		rz_vector_fini(&type->struct_data.members);
+		break;
+	case RZ_BASE_TYPE_KIND_UNION:
+		rz_vector_fini(&type->union_data.members);
+		break;
+	case RZ_BASE_TYPE_KIND_ENUM:
+		rz_vector_fini(&type->enum_data.cases);
+		break;
+	case RZ_BASE_TYPE_KIND_TYPEDEF:
+	case RZ_BASE_TYPE_KIND_ATOMIC:
+		break;
+	default:
+		break;
+	}
+	RZ_FREE(type);
+}
+
+RZ_API RzBaseType *rz_type_base_type_new(RzBaseTypeKind kind) {
+	RzBaseType *type = RZ_NEW0(RzBaseType);
+	if (!type) {
+		return NULL;
+	}
+	type->kind = kind;
+	switch (type->kind) {
+	case RZ_BASE_TYPE_KIND_STRUCT:
+		rz_vector_init(&type->struct_data.members, sizeof(RzTypeStructMember), rz_type_base_struct_member_free, NULL);
+		break;
+	case RZ_BASE_TYPE_KIND_ENUM:
+		rz_vector_init(&type->enum_data.cases, sizeof(RzTypeEnumCase), rz_type_base_enum_case_free, NULL);
+		break;
+	case RZ_BASE_TYPE_KIND_UNION:
+		rz_vector_init(&type->union_data.members, sizeof(RzTypeUnionMember), rz_type_base_union_member_free, NULL);
+		break;
+	default:
+		break;
+	}
+
+	return type;
+}
+
+/**
+ * @brief Saves RzBaseType into the SDB
+ *
+ * @param t
+ * @param type RzBaseType to save
+ * @param name Name of the type
+ */
+RZ_API void rz_type_save_base_type(const RzType *t, const RzBaseType *type) {
+	rz_return_if_fail(t && type && type->name);
+
+	// TODO, solve collisions, if there are 2 types with the same name and kind
+
+	switch (type->kind) {
+	case RZ_BASE_TYPE_KIND_STRUCT:
+		save_struct(t, type);
+		break;
+	case RZ_BASE_TYPE_KIND_ENUM:
+		save_enum(t, type);
+		break;
+	case RZ_BASE_TYPE_KIND_UNION:
+		save_union(t, type);
+		break;
+	case RZ_BASE_TYPE_KIND_TYPEDEF:
+		save_typedef(t, type);
+		break;
+	case RZ_BASE_TYPE_KIND_ATOMIC:
+		save_atomic_type(t, type);
+		break;
+	default:
+		break;
+	}
+}
+
