+++ librz/type/code.c
@@ -1,9 +1,9 @@
 // SPDX-FileCopyrightText: 2013-2019 pancake <pancake@nopcode.org>
 // SPDX-License-Identifier: LGPL-3.0-only
 
-#include "rz_util.h"
-#include "rz_types.h"
-#include "rz_parse.h"
+#include <rz_util.h>
+#include <rz_type.h>
+
 #include "tcc.h"
 
 extern int tcc_sym_push(char *typename, int typesize, int meta);
@@ -31,7 +31,7 @@ static bool __typeLoad(void *p, const char *k, const char *v) {
 		return false;
 	}
 	int btype = 0;
-	RzAnalysis *analysis = (RzAnalysis *)p;
+	RzType *t = (RzType *)p;
 	//rz_cons_printf ("tk %s=%s\n", k, v);
 	// TODO: Add unions support
 	if (!strncmp(v, "struct", 6) && strncmp(k, "struct.", 7)) {
@@ -41,7 +41,7 @@ static bool __typeLoad(void *p, const char *k, const char *v) {
 		int typesize = 0;
 		// TODO: Add typesize here
 		char *query = sdb_fmt("struct.%s", k);
-		char *members = sdb_get(analysis->sdb_types, query, 0);
+		char *members = sdb_get(t->sdb_types, query, 0);
 		char *next, *ptr = members;
 		if (members) {
 			do {
@@ -50,7 +50,7 @@ static bool __typeLoad(void *p, const char *k, const char *v) {
 					break;
 				}
 				query = sdb_fmt("struct.%s.%s", k, name);
-				char *subtype = sdb_get(analysis->sdb_types, query, 0);
+				char *subtype = sdb_get(t->sdb_types, query, 0);
 				if (!subtype) {
 					break;
 				}
@@ -64,7 +64,7 @@ static bool __typeLoad(void *p, const char *k, const char *v) {
 					char *subname = tmp;
 					// TODO: Go recurse here
 					query = sdb_fmt("struct.%s.%s.meta", subtype, subname);
-					btype = sdb_num_get(analysis->sdb_types, query, 0);
+					btype = sdb_num_get(t->sdb_types, query, 0);
 					tcc_sym_push(subtype, 0, btype);
 				}
 				free(subtype);
@@ -94,15 +94,15 @@ static void __errorFunc(void *opaque, const char *msg) {
 	}
 }
 
-RZ_API char *rz_parse_c_file(RzAnalysis *analysis, const char *path, const char *dir, char **error_msg) {
+RZ_API char *rz_type_parse_c_file(RzType *t, const char *path, const char *dir, char **error_msg) {
 	char *str = NULL;
-	TCCState *T = tcc_new(analysis->cpu, analysis->bits, analysis->os);
+	TCCState *T = tcc_new(t->target->cpu, t->target->bits, t->target->os);
 	if (!T) {
 		return NULL;
 	}
 	tcc_set_callback(T, &__appendString, &str);
 	tcc_set_error_func(T, (void *)error_msg, __errorFunc);
-	sdb_foreach(analysis->sdb_types, __typeLoad, analysis);
+	sdb_foreach(t->sdb_types, __typeLoad, t);
 	if (tcc_add_file(T, path, dir) == -1) {
 		free(str);
 		str = NULL;
@@ -111,15 +111,15 @@ RZ_API char *rz_parse_c_file(RzAnalysis *analysis, const char *path, const char 
 	return str;
 }
 
-RZ_API char *rz_parse_c_string(RzAnalysis *analysis, const char *code, char **error_msg) {
+RZ_API char *rz_type_parse_c_string(RzType *t, const char *code, char **error_msg) {
 	char *str = NULL;
-	TCCState *T = tcc_new(analysis->cpu, analysis->bits, analysis->os);
+	TCCState *T = tcc_new(t->target->cpu, t->target->bits, t->target->os);
 	if (!T) {
 		return NULL;
 	}
 	tcc_set_callback(T, &__appendString, &str);
 	tcc_set_error_func(T, (void *)error_msg, __errorFunc);
-	sdb_foreach(analysis->sdb_types, __typeLoad, NULL);
+	sdb_foreach(t->sdb_types, __typeLoad, NULL);
 	if (tcc_compile_string(T, code) != 0) {
 		free(str);
 		str = NULL;
@@ -129,6 +129,6 @@ RZ_API char *rz_parse_c_string(RzAnalysis *analysis, const char *code, char **er
 }
 
 // XXX do not use globals
-RZ_API void rz_parse_c_reset(RzParse *p) {
+RZ_API void rz_type_parse_c_reset(RzType *p) {
 	anon_sym = SYM_FIRST_ANOM;
 }
