+++ librz/util/meson.build
@@ -31,7 +31,6 @@ rz_util_sources = [
   'log.c',
   'mem.c',
   'name.c',
-  'format.c',
   'pie.c',
   'seven.c',
   'print.c',
@@ -43,7 +42,6 @@ rz_util_sources = [
   'astr.c',
   'pkcs7.c',
   'x509.c',
-  'utype.c',
   'randomart.c',
   'range.c',
   'rbtree.c',
