+++ test/unit/meson.build
@@ -11,7 +11,6 @@ if get_option('enable_tests')
     'analysis_function',
     'analysis_hints',
     'analysis_meta',
-    'analysis_types',
     'analysis_var',
     'analysis_xrefs',
     'annotated_code',
@@ -53,7 +52,6 @@ if get_option('enable_tests')
     'json',
     'list',
     'ovf',
-    'parse_ctype',
     'pdb',
     'pj',
     'queue',
@@ -65,6 +63,7 @@ if get_option('enable_tests')
     'serialize_config',
     'serialize_flag',
     'serialize_spaces',
+    'serialize_types',
     'sign',
     'skiplist',
     'skyline',
@@ -77,6 +76,7 @@ if get_option('enable_tests')
     'table',
     'task',
     'tree',
+	'type',
     'uleb128',
     'unum',
     'util',
@@ -106,6 +106,7 @@ if get_option('enable_tests')
         rz_bp_dep,
         rz_reg_dep,
         rz_syscall_dep,
+		rz_type_dep,
         rz_analysis_dep,
         rz_parse_dep,
         rz_egg_dep,
